requirements.txt
django==4.2.8
djangorestframework
requests
django-cors-headers
